seek based on cbr
