+++ packages/media-parser/src/test/seeking/seek-mp3-cbr.test.ts
@@ -1,0 +1,54 @@
+import {exampleVideos} from '@remotion/example-videos';
+import {expect, test} from 'bun:test';
+import {mediaParserController} from '../../controller/media-parser-controller';
+import {nodeReader} from '../../node';
+import {parseMedia} from '../../parse-media';
+
+test('seek mp3 cbr', async () => {
+	const controller = mediaParserController();
+	controller._experimentalSeek({
+		type: 'keyframe-before-time',
+		timeInSeconds: 60,
+	});
+
+	let samples = 0;
+
+	await parseMedia({
+		src: exampleVideos.music,
+		acknowledgeRemotionLicense: true,
+		controller,
+		reader: nodeReader,
+		onAudioTrack: () => {
+			return (sample) => {
+				samples++;
+				if (samples === 1) {
+					expect(sample.timestamp / sample.timescale).toBe(59.95102);
+					controller._experimentalSeek({
+						type: 'keyframe-before-time',
+						timeInSeconds: 10000000,
+					});
+				}
+
+				if (samples === 2) {
+					expect(sample.timestamp / sample.timescale).toBe(125.048163);
+				}
+
+				if (samples === 3) {
+					throw new Error(
+						'Should not reach this, this should have been the last sample',
+					);
+				}
+			};
+		},
+	});
+
+	expect(samples).toBe(2);
+
+	const [firstSeek] =
+		controller._internals.performedSeeksSignal.getPerformedSeeks();
+	expect(firstSeek).toEqual({
+		from: 5141,
+		to: 2401120,
+		type: 'user-initiated',
+	});
+});
